some changes

+++ cogs/_staff.py
@@ -12,7 +12,7 @@ class _staff(commands.Cog):
     description="Adds a card to database. Staff only command", guild_ids=[1024399481782927511, 1024402764601765978])
     async def add_card(self, inter, name: str, rarity: commands.Range[1, 5], group: str, id,
                         pic: discord.Attachment, limited_days:int=None): #type:ignore
-        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and inter.author.id != 756018524413100114:
+        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and not inter.guild.get_role(1024979194800779324) in inter.author.roles and inter.author.id != 756018524413100114:
             return await inter.send("You don't have permissions to use this command!")
         await inter.response.defer()
         await inter.edit_original_message(
@@ -25,7 +25,7 @@ class _staff(commands.Cog):
     
     @commands.slash_command(description="Makes a card not accessible. Staff only command", guild_ids=[1024399481782927511, 1024402764601765978])
     async def delete_card(self, inter, id: str):
-        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and inter.author.id != 756018524413100114:
+        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and not inter.guild.get_role(1024979194800779324) in inter.author.roles and inter.author.id != 756018524413100114:
             return await inter.send("You don't have permissions to use this command!")
         await inter.send("Succesfully deleted")
         await self.bot.delete_card(id, from_existance=False)
@@ -33,6 +33,8 @@ class _staff(commands.Cog):
 
     @commands.slash_command(description="Takes a card from a user. Staff only command", guild_ids=[1024399481782927511, 1024402764601765978])
     async def take_card(self, inter, user: discord.User):
+        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and not inter.guild.get_role(1024979194800779324) in inter.author.roles and inter.author.id != 756018524413100114:
+            return await inter.send("You don't have permissions to use this command!")
         r = await self.bot.get_inventory(user.id)
         view = DeleteView(user.id, r)
         view.bot = self.bot #type:ignore
@@ -42,7 +44,7 @@ class _staff(commands.Cog):
     
     @commands.slash_command(description="Gives a card to anyone. Staff only command", guild_ids=[1024399481782927511, 1024402764601765978])
     async def give_card(self, inter, user: discord.User, card_id:str):
-        if not inter.guild.get_role(1024979194800779324) in inter.author.roles:
+        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and not inter.guild.get_role(1024979194800779324) in inter.author.roles and inter.author.id != 756018524413100114:
             return await inter.send("You don't have permissions to use this command!")
         if card_id == "Nothing found":
             await inter.send("You don't have any card in your inventory.", ephemeral=True)
@@ -58,7 +60,7 @@ class _staff(commands.Cog):
 
     @commands.slash_command(description="Removes a card from database. Staff only command", guild_ids=[1024399481782927511, 1024402764601765978])
     async def remove_card(self, inter, card_id):
-        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and inter.author.id != 756018524413100114:
+        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and not inter.guild.get_role(1024979194800779324) in inter.author.roles and inter.author.id != 756018524413100114:
             return await inter.send("You don't have permissions to use this command!")
         await inter.send("Succesfully deleted")
         await self.bot.delete_card(card_id, from_existance=True)
@@ -67,6 +69,14 @@ class _staff(commands.Cog):
     @remove_card.autocomplete("card_id")
     async def remgetallids(self, inter, user_input):
         return [id for id in self.bot.data.keys() if user_input.lower() in id.lower()][:25]
+    
+
+    @commands.slash_command(description="Give petals to anyone. Staff only command", guild_ids=[1024399481782927511, 1024402764601765978])
+    async def give_petals(self, inter, user:discord.User, petals: int):
+        if not inter.guild.get_role(1024579633699627028) in inter.author.roles and not inter.guild.get_role(1024979194800779324) in inter.author.roles and inter.author.id != 756018524413100114:
+            return await inter.send("You don't have permissions to use this command!")
+        await self.bot.add_coins(user.id, petals)
+
 
 
 def setup(bot: Hanknyeon):
